potential_file can sometimes be None!

+++ librarian_background/corruption_fixer.py
@@ -179,7 +179,8 @@ class CorruptionFixer(Task):
             # This also deletes remote instances which will need to be repaired. However
             # it is unlikely that we will be in that situation. Unfortunately we _must_ commit
             # this as the files table must be accessed from a different table.
-            potential_file.delete(session=session, commit=True, force=True)
+            if potential_file is not None:
+                potential_file.delete(session=session, commit=True, force=True)
 
             resend_request = CorruptionResendRequest(
                 file_name=corrupt.file_name,
